+++ examples/strategy_backtest.py
@@ -2,17 +2,17 @@ import click
 
 from decimal import Decimal
 
-from qstrader.backtest.backtest import Backtest
-from qstrader.execution_handler.execution_handler import IBSimulatedExecutionHandler
-from qstrader.portfolio_handler.portfolio_handler import PortfolioHandler
-from qstrader.position_sizer.position_sizer import TestPositionSizer
-from qstrader.price_handler.price_handler import HistoricCSVPriceHandler
-from qstrader.risk_manager.risk_manager import TestRiskManager
-from qstrader.statistics.statistics import SimpleStatistics
-from qstrader.compliance.compliance import TestCompliance
 from qstrader import settings
-from qstrader.strategy.strategy import TestStrategy
-from qstrader.compat.compat import queue
+from qstrader.compat import queue
+from qstrader.price_handler.historic_csv_tick import HistoricCSVTickPriceHandler
+from qstrader.strategy.example import ExampleStrategy
+from qstrader.position_sizer.fixed import FixedPositionSizer
+from qstrader.risk_manager.example import ExampleRiskManager
+from qstrader.portfolio_handler import PortfolioHandler
+from qstrader.compliance.csv import CsvCompliance
+from qstrader.execution_handler.ib_simulated import IBSimulatedExecutionHandler
+from qstrader.statistics.simple import SimpleStatistics
+from qstrader.trading_session.backtest import Backtest
 
 
 def run(config, testing):
@@ -26,18 +26,18 @@ def run(config, testing):
     # max_iters = 10000000000
 
     # Use Historic CSV Price Handler
-    price_handler = HistoricCSVPriceHandler(
+    price_handler = HistoricCSVTickPriceHandler(
         csv_dir, events_queue, tickers
     )
 
-    # Use the Test Strategy
-    strategy = TestStrategy(tickers, events_queue)
+    # Use the Example Strategy
+    strategy = ExampleStrategy(tickers, events_queue)
 
     # Use an example Position Sizer
-    position_sizer = TestPositionSizer()
+    position_sizer = FixedPositionSizer()
 
     # Use an example Risk Manager
-    risk_manager = TestRiskManager()
+    risk_manager = ExampleRiskManager()
 
     # Use the default Portfolio Handler
     portfolio_handler = PortfolioHandler(
@@ -45,8 +45,8 @@ def run(config, testing):
         position_sizer, risk_manager
     )
 
-    # Use the TestCompliance component
-    compliance = TestCompliance(config)
+    # Use the CsvCompliance component
+    compliance = CsvCompliance(config)
 
     # Use a simulated IB Execution Handler
     execution_handler = IBSimulatedExecutionHandler(
