+++ qstrader/compat.py
