+++ qstrader/compliance/csv.py
@@ -1,37 +1,11 @@
-from abc import ABCMeta, abstractmethod
-
 import datetime
 import os
 import csv
 
-
-class Compliance(object):
-    """
-    The Compliance component should be given every trade
-    that occurs in qstrader.
-
-    It is designed to keep track of anything that may
-    be required for regulatory or audit (or debugging)
-    purposes. Extended versions can write trades to a
-    CSV, or a database.
-    """
-
-    __metaclass__ = ABCMeta
-
-    @abstractmethod
-    def record_trade(self, fill):
-        """
-        Takes a FillEvent from an ExecutionHandler
-        and logs each of these.
-
-        Parameters:
-        fill - A FillEvent with information about the
-        trade that has just been executed.
-        """
-        raise NotImplementedError("Should implement record_trade()")
+from .base import AbstractCompliance
 
 
-class TestCompliance(Compliance):
+class CsvCompliance(AbstractCompliance):
     """
     A basic compliance module which writes trades to a
     CSV file in the output directory.
@@ -49,7 +23,6 @@ class TestCompliance(Compliance):
         self.config = config
         # Remove the previous CSV file
         today = datetime.datetime.utcnow().date()
-        self.csv_filename = "tradelog_" + datetime.datetime.today().strftime("%Y-%m-%d") + ".csv"
         self.csv_filename = "tradelog_" + today.strftime("%Y-%m-%d") + ".csv"
 
         try:
