+++ qstrader/execution_handler/ib_simulated.py
@@ -1,42 +1,10 @@
 from decimal import Decimal
-from abc import ABCMeta, abstractmethod
 
-from qstrader.event.event import (FillEvent, EventType)
-from qstrader.price_handler.price_handler import PriceHandlerType
-
-
-class ExecutionHandler(object):
-    """
-    The ExecutionHandler abstract class handles the interaction
-    between a set of order objects generated by a PortfolioHandler
-    and the set of Fill objects that actually occur in the
-    market.
-
-    The handlers can be used to subclass simulated brokerages
-    or live brokerages, with identical interfaces. This allows
-    strategies to be backtested in a very similar manner to the
-    live trading engine.
-
-    ExecutionHandler can link to an optional Compliance component
-    for simple record-keeping, which will keep track of all executed
-    orders.
-    """
-
-    __metaclass__ = ABCMeta
-
-    @abstractmethod
-    def execute_order(self, event):
-        """
-        Takes an OrderEvent and executes it, producing
-        a FillEvent that gets placed onto the events queue.
-
-        Parameters:
-        event - Contains an Event object with order information.
-        """
-        raise NotImplementedError("Should implement execute_order()")
+from .base import AbstractExecutionHandler
+from ..event import (FillEvent, EventType)
 
 
-class IBSimulatedExecutionHandler(ExecutionHandler):
+class IBSimulatedExecutionHandler(AbstractExecutionHandler):
     """
     The simulated execution handler for Interactive Brokers
     converts all order objects into their equivalent fill
@@ -84,7 +52,7 @@ class IBSimulatedExecutionHandler(ExecutionHandler):
             quantity = event.quantity
 
             # Obtain the fill price
-            if self.price_handler.type == PriceHandlerType.TICK:
+            if self.price_handler.istick():
                 bid, ask = self.price_handler.get_best_bid_ask(ticker)
                 if event.action == "BOT":
                     fill_price = Decimal(str(ask))
