--- qstrader/price_handler/__init__.py
@@ -1,4 +1,0 @@
-from enum import Enum
-
-
-PriceHandlerType = Enum("PriceHandlerType", "TICK BAR")
